refactor(#174): 채팅창 드래그 관련 함수 useDrag 커스텀 훅으로 분리
